AI_ChatBot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.354
langchain-community==0.0.13
langchain-core==0.1.3
langchain-groq==0.0.3
langchain-huggingface==0.0.2
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.2